updating subnet test cases

+++ src/test/java/io/dchq/sdk/core/subnet/SubnetTest.java
@@ -35,10 +35,16 @@ public class SubnetTest extends AbstractServiceTest {
 		ResponseEntity<List<NetworkACL>> response = networkACLService.findAll();
 		if (response.getResults() != null && !response.isErrors()) {
 			for (NetworkACL acl : response.getResults()) {
-				logger.info("cleaning up Network ACL with name - "+acl.getName());
-				ResponseEntity<?> deleteResponse = networkACLService.delete(acl.getId());
-				for (Message message : deleteResponse.getMessages()) {
-					logger.warn("Error Network ACL deletion: [{}] ", message.getMessageText());
+				
+				NetworkACL netowrkACL = networkACLService.findById(acl.getId()).getResults();
+				
+				if(netowrkACL.getSubnet().getId().equals(this.subnetCreated.getId())){
+					logger.info("cleaning up Network ACL with name - "+acl.getName());
+					
+					ResponseEntity<?> deleteResponse = networkACLService.delete(acl.getId());
+					for (Message message : deleteResponse.getMessages()) {
+						logger.warn("Error Network ACL deletion: [{}] ", message.getMessageText());
+					}
 				}
 			}
 		}
@@ -48,10 +54,16 @@ public class SubnetTest extends AbstractServiceTest {
 		ResponseEntity<List<SecurityGroup>> response = securityGroupService.findAll();
 		if (response.getResults() != null && !response.isErrors()) {
 			for (SecurityGroup sg : response.getResults()) {
-				logger.info("cleaning up Security Group with name - "+sg.getName());
-				ResponseEntity<?> deleteResponse = securityGroupService.delete(sg.getId());
-				for (Message message : deleteResponse.getMessages()) {
-					logger.warn("Error Security Group deletion: [{}] ", message.getMessageText());
+				
+				SecurityGroup securityGroup = securityGroupService.findById(sg.getId()).getResults();
+				
+				if(securityGroup.getSubnet().getId().equals(this.subnetCreated.getId())){
+					logger.info("cleaning up Security Group with name - "+sg.getName());
+					
+					ResponseEntity<?> deleteResponse = securityGroupService.delete(sg.getId());
+					for (Message message : deleteResponse.getMessages()) {
+						logger.warn("Error Security Group deletion: [{}] ", message.getMessageText());
+					}
 				}
 			}
 		}
